Update: Modify join key max size to 8 from 11.

+++ source/bzs/db/protocol/tdap/client/fields.h
@@ -189,7 +189,7 @@ public:
 
     inline field operator[](const std::_tstring& name) const
     {
-        int index = m_fns->indexByName(name);
+        short index = m_fns->indexByName(name);
         return operator[](index);
     }
